requirements.txt
asarPy==1.0.1
asgiref==3.8.1
Django==5.1.5
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
